requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
jinja2==3.1.4
telethon==1.36.0
python-dateutil==2.9.0.post0
pydantic==2.9.2
PyYAML==6.0.2
